// src/app/interfaces/AprobacionAutoridades.ts
export interface DataAprobacionAutoridades {
    resultadoAprobacionPresidenteMunicipal: string;
    resultadoAprobacionGobernadorEstado: string;
    resultadoAprobacionPresidenteMexico: string;
    resultadoEconomicoPresidenteMunicipal: string;
    resultadoEconomicoGobernadorEstado: string;
    resultadoEconomicoPresidenteMexico: string;
    resultadoInseguridadPresidenteMunicipal: string;
    resultadoInseguridadGobernadorEstado: string;
    resultadoInseguridadPresidenteMexico: string;
    idPrincipalProblemaMunicipio: string;
    idPrincipalProblemaEstado : string;
    idPrincipalProblemaPais: string;
}